+++ src/components/Breadcrumbs/breadcrumbsChangelogData.ts
@@ -13,6 +13,13 @@ export const changelogData: ChangelogData[] = [
     date: "2025-01-30",
     version: "3.5.3",
     type: "Update",
+    affects: ["Styles"],
+    notes: ["Updates arrow to be center aligned vertically."],
+  },
+  {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
     affects: ["Accessibility"],
     notes: [
       "Adds screenreaderOnly styles to span inside breadcrumb so link has descriptive text on mobile.",