+++ CHANGELOG.md
@@ -18,7 +18,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates the `Pagination` component to handle 4 digit page counts.
 - Updates the `Heading` component to add `line-height` styles for mobile.
-- Updates the `Breadcrumbs` component to make link text screen readable on mobile breakpoints.
+- Updates the `Breadcrumbs` component to make link text screen readable on mobile breakpoints
+- Updates the `Breadcrumbs` component to center align arrow on mobile breakpoints
 - Updates the `MultiSelect` component to adjust the text and size styles for the search input field.
 
 ## 3.5.2 (January 16, 2025)