+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the `StatusBadge` component to support the `fontSize` style prop.
+
 ## 3.5.3 (January 30, 2025)
 
 ### Adds