+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates the `StatusBadge` component to support the `fontSize` style prop.
 - Updates the `Checkbox` component to set the correct background color for all states.
 
 ## 3.5.3 (January 30, 2025)