+++ CHANGELOG.md
@@ -12,7 +12,12 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 - Updates the `StatusBadge` component to support the `fontSize` style prop.
 - Updates the `Checkbox` component to set the correct background color for all states.
+- Updates the `Menu` component to set the line height to 1.5 for menu options.
 - Updates the `FeedbackBox` component to set the font weight for the popup header to `medium`.
+
+### Fixes
+
+- Fixes the circular styles for images in the `Menu` component.
 
 ## 3.5.3 (January 30, 2025)