+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Checkbox` component to set the correct background color for all states.
 - Updates the `Menu` component to set the line height to 1.5 for menu options.
 - Updates the `FeedbackBox` component to set the font weight for the popup header to `medium`.
+- Updates the `primary` variant of the `Hero` component to have appropriate padding for desktop and mobile, and `max-width` of 860px
+- Updates the `campaign` variant of the `Hero` component to have consistent padding.
 
 ### Fixes