+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds the `useResponsiveSpacing` hook.
 - Adds docs for `Responsive Grid` Style Guide.
 
 ### Updates