+++ CHANGELOG.md
@@ -10,14 +10,19 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds the `"utilityCheck"` option to the `Icon` component.
 - Adds the `getTextFromElement` utility function.
 
 ### Updates
 
+- Updates the `Checkbox` component to use the DS `utilityCheck` icon.
+- Updates the `AlphabetFilter` component to correct the font weight.
 - Updates the `StatusBadge` component to support the `fontSize` style prop.
 - Updates the `Checkbox` component to set the correct background color for all states.
 - Updates the `Menu` component to set the line height to 1.5 for menu options.
 - Updates the `FeedbackBox` component to set the font weight for the popup header to `medium`.
+- Updates the `primary` variant of the `Hero` component to have appropriate padding for desktop and mobile, and `max-width` of 860px
+- Updates the `campaign` variant of the `Hero` component to have consistent padding.
 - Updates the `Label` component to set the font weight for the `(required)` label to `"medium"`.
 - Updates the `Fieldset` component to set the font weight for the `(required)` label to `"medium"`.
 - Updates the `FeedbackBox` component to set the font weight for the `(required)` label on the `Textarea` element to `"medium"`.